refactor(context): reorganize GlobalContext import paths and create GlobalContext file
